agents: files, agents and google namespaces run on the capability registry (PR 9 of tool-class retirement)
Ports the workspace file set, todo_write, run_subtask/run_search (the
capability constructs the existing SubAgentTool classes over
run.subAgent, machinery untouched), and the twenty Google Workspace
tools. The todo store moves beside its writer in capabilities/files.ts.
