+++ packages/agents/tests/capabilities-registry.test.ts
@@ -38,7 +38,40 @@ const CAPABILITY_CATEGORY_SNAPSHOT: Record<string, PermissionCategory> = {
   // default classes the three judges and the taste pair `external`.
   compare_images: "external",
   create_workflow: "write",
-  critique_image: "external",
+  critique_image: "external",,
+  edit_file: "write",
+  glob: "read",
+  grep: "read",
+  list_directory: "read",
+  read_file: "read",
+  // Both spawn a child loop whose own tools are gated inside it, so the
+  // spawning call itself has no side effect.
+  run_search: "read",
+  run_subtask: "read",
+  todo_write: "read",
+  write_file: "write",
+  // Google Workspace: none is listed in `TOOL_PERMISSION_CATEGORIES`, so the
+  // map's conservative default classes them `external`. Carried over as-is.
+  gmail_get_message: "external",
+  gmail_list_labels: "external",
+  gmail_modify_labels: "external",
+  gmail_search: "external",
+  gmail_send_message: "external",
+  google_calendar_create_event: "external",
+  google_calendar_delete_event: "external",
+  google_calendar_list_calendars: "external",
+  google_calendar_list_events: "external",
+  google_docs_append: "external",
+  google_docs_create: "external",
+  google_docs_read: "external",
+  google_drive_create_file: "external",
+  google_drive_get_file: "external",
+  google_drive_read_file: "external",
+  google_drive_search: "external",
+  google_sheets_append: "external",
+  google_sheets_create: "external",
+  google_sheets_read: "external",
+  google_sheets_update: "external"
   debug_workflow: "execute",
   edit_image: "write",
   embed_text: "write",
